requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1  
pandas==2.2.1  
requests==2.31.0  
shutilwhich==1.1.0  
